+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.6.0] - 2020-01-06
 
 ### Adds
 - Adds a `dropdown_content` function to the `AuthProvider`.
+- Adds an `Array` field
 
 ### Changes
 - [breaking] Changes `gravatar` to `avatar_url` for the current_user map